Move CLI logic to `cli/main.rb`
