+++ exe/tapioca
@@ -18,4 +18,4 @@ end
 
 require_relative "../lib/tapioca/internal"
 
-Tapioca::Cli.start(ARGV)
+Tapioca::Cli::Main.start(ARGV)
