+++ lib/tapioca/internal.rb
@@ -10,6 +10,7 @@ require "tapioca/config"
 require "tapioca/config_builder"
 require "tapioca/generator"
 require "tapioca/cli"
+require "tapioca/cli/main"
 require "tapioca/gemfile"
 require "tapioca/compilers/sorbet"
 require "tapioca/compilers/requires_compiler"
